Do not store inputIndex and GrammarSlot in GSSNode
