+++ src/org/iguana/gss/GSSEdge.java
@@ -1,69 +1,19 @@
-/*
- * Copyright (c) 2015, Ali Afroozeh and Anastasia Izmaylova, Centrum Wiskunde & Informatica (CWI)
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without 
- * modification, are permitted provided that the following conditions are met:
- *
- * 1. Redistributions of source code must retain the above copyright notice, this 
- *    list of conditions and the following disclaimer.
- *
- * 2. Redistributions in binary form must reproduce the above copyright notice, this 
- *    list of conditions and the following disclaimer in the documentation and/or 
- *    other materials provided with the distribution.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND 
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED 
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. 
- * IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, 
- * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT 
- * NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, 
- * OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, 
- * WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) 
- * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY 
- * OF SUCH DAMAGE.
- *
- */
-
 package org.iguana.gss;
 
 import org.iguana.datadependent.env.Environment;
 import org.iguana.grammar.slot.BodyGrammarSlot;
 import org.iguana.result.Result;
 
-public class GSSEdge<T extends Result> {
+public interface GSSEdge<T extends Result> {
 
-	private final BodyGrammarSlot returnSlot;
-	private final T result;
-	private final GSSNode<T> destination;
-	private final Environment env;
+    T getResult();
 
-	GSSEdge(BodyGrammarSlot slot, T result, GSSNode<T> destination, Environment env) {
-		this.returnSlot = slot;
-		this.result = result;
-		this.destination = destination;
-		this.env = env;
-	}
+    int getInputIndex();
 
-	public T getResult() {
-		return result;
-	}
+    BodyGrammarSlot getReturnSlot();
 
-	public BodyGrammarSlot getReturnSlot() {
-		return returnSlot;
-	}
+    GSSNode<T> getDestination();
 
-	public GSSNode<T> getDestination() {
-		return destination;
-	}
-
-    public Environment getEnv() {
-        return env;
-    }
-
-    @Override
-	public String toString() {
-		return String.format("(%s, %s, %s)", returnSlot, result, destination);
-	}
+    Environment getEnv();
 
 }
