+++ src/org/iguana/result/Result.java
@@ -2,6 +2,7 @@ package org.iguana.result;
 
 public interface Result {
     int getIndex();
+    int getLeftExtent();
     boolean isDummy();
     Object getValue();
 }
